requirements.txt
absl-py==1.2.0
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==4.1.0
blinker==1.6.2
cachetools==5.2.0
celery==5.3.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.22.4
colorama==0.4.4
colorlover==0.3.0
comtypes==1.1.13
cufflinks==0.17.3
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.3
decorator==5.1.1
deepface==0.0.75
distlib==0.3.6
Django==4.1.3
django-crispy-forms==1.14.0
dlib==19.24.0
docopt==0.6.2
entrypoints==0.4
executing==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.8.0
fire==0.4.0
Flask==2.2.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
flatbuffers==22.10.26
fonttools==4.34.4
gast==0.4.0
gdown==4.5.3
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.50.0
h5py==3.7.0
idna==3.3
imutils==0.5.4
ipykernel==6.15.3
ipython==8.5.0
ipywidgets==8.0.3
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.4
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.4.4
kombu==5.3.1
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mediapipe==0.8.10.1
mtcnn==0.1.1
mysqlclient==2.1.1
nest-asyncio==1.5.5
numpy==1.22.4
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
pipreqs==0.4.11
platformdirs==3.1.1
plotly==5.11.0
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
redis==4.6.0
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.8.1
seaborn==0.12.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.10
sqlparse==0.4.3
stack-data==0.5.1
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
tflearn==0.5.0
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.4
windows-curses==2.3.1
wrapt==1.14.1
yarg==0.1.9
